requirements.txt
anthropic>=0.40.0
openpyxl>=3.1.0
pandas>=2.0.0
chromadb>=0.5.0
sentence-transformers>=3.0.0
python-dotenv>=1.0.0
rich>=13.7.0
